Disable intraline diff highlighting algorithm for lines which contain UTF-8 text
Summary: This is sort of cheating, but just have this feature disable itself if the input contains multibyte UTF-8 characters. We can clean it up in the future, maybe when we have better utf8 tools.

This means that all the algorithms are safe to pass utf8 to, so we can get rid of all the "<?>" silliness.

Test Plan: Added a UTF8 character to a line, diffed it out, and got the entire line highlighted as changed. See: https://secure.phabricator.com/file/view/PHID-FILE-70fb54eb3f88dc057ab3/

Reviewers: jungejason, aran, tuomaspelkonen

CC:

Differential Revision: 514